Moved signal handler assignment in SetUpSignal
