+++ signalHandler.c
@@ -1,4 +1,5 @@
 #include "signalHandler.h"
+#include "executoner.h"
 #include <stdio.h>
 #include <signal.h>
 
@@ -12,6 +13,7 @@ void CatchSignal(int signalNumber)
 void SetUpSignal(int* validCommand)
 {
         setValidCommand = validCommand;
+
         struct sigaction act;
 	act.sa_flags = 0;
 	act.sa_handler = CatchSignal;
@@ -19,4 +21,12 @@ void SetUpSignal(int* validCommand)
 	//sigaction(SIGINT, &act, NULL);
 	sigaction(SIGQUIT, &act, NULL);
 	sigaction(SIGKILL, &act, NULL);
+
+	// Set up handler for SIGCHLD signal (see executoner.c)
+  	struct sigaction childDone;
+		
+  	childDone.sa_flags = SA_RESTART;
+  	childDone.sa_flags |= SA_SIGINFO;
+  	childDone.sa_sigaction = ChildHandler;
+  	sigaction(SIGCHLD,&childDone,NULL);
 }
